pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Gamabyta24 <Gamabyta@gmail.com>"]
readme = "README.md"
packages = [
    { include = "gendiff" },
    { include = "tests" },
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
codeclimate-test-reporter = "^0.2.3"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
